Inputs login data into sql table
It takes username, password, first name, and last name as input. The database has a UserID column that needs data as well, so either we should manually throw in a bunch of numbers 1-100 or however many users we want, or implement some way of check the database for the last UserID and incrementing that by one for the next UserID. 

We could also make a feature to delete a user and their subsequent data from the database as an admin privilege, an implementation of the banning feature. This is not a necessary feature.